main을 전부  fragment로 이동
