+++ NackUn/StudyFirst/app/src/main/java/sample/nackun/com/studyfirst/main/MainFragment.kt
@@ -2,17 +2,17 @@ package sample.nackun.com.studyfirst.main
 
 import android.os.Bundle
 import android.support.v4.app.Fragment
-import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import kotlinx.android.synthetic.main.main_recyclerview.view.*
+import android.widget.TextView
+import android.widget.Toast
+import android.widget.Toast.LENGTH_SHORT
+import kotlinx.android.synthetic.main.main_fragment.view.*
 import sample.nackun.com.studyfirst.R
 import sample.nackun.com.studyfirst.TickerAdapter
 import sample.nackun.com.studyfirst.data.Repository
 import sample.nackun.com.studyfirst.data.remote.RemoteDataSource
-import sample.nackun.com.studyfirst.vo.Market
-import sample.nackun.com.studyfirst.vo.Ticker
 
 class MainFragment : Fragment(), MainContract.View{
     override var marketName: String= "BTC"
@@ -21,25 +21,36 @@ class MainFragment : Fragment(), MainContract.View{
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
                               savedInstanceState: Bundle?): View? {
-        val root = inflater.inflate(R.layout.main_recyclerview, container, false)
+        val root = inflater.inflate(R.layout.main_fragment, container, false)
 
         with(root){
+            val onClickListener = object : View.OnClickListener {
+                override fun onClick(v: View?) {
+                    marketKRW.setTextColor(resources.getColor(R.color.grey))
+                    marketBTC.setTextColor(resources.getColor(R.color.grey))
+                    marketETH.setTextColor(resources.getColor(R.color.grey))
+                    marketUSDT.setTextColor(resources.getColor(R.color.grey))
+                    var selectedMarket = v as TextView
+                    selectedMarket.setTextColor(resources.getColor(R.color.indigo))
+                    marketName = selectedMarket.text.toString()
+                    selectedMarket(marketName)
+                }
+            }
+            marketKRW.setOnClickListener(onClickListener)
+            marketBTC.setOnClickListener(onClickListener)
+            marketETH.setOnClickListener(onClickListener)
+            marketUSDT.setOnClickListener(onClickListener)
+
             tickerRecyclerView.adapter =  tickerAdapter
-            //val recyclerView = findViewById<RecyclerView>(R.id.tickerRecyclerView).let { it.adapter =  tickerAdapter}
+            marketKRW.callOnClick()
         }
         val mainPresenter = MainPresenter(Repository(RemoteDataSource), this)
         mainPresenter.start()
-        mainPresenter.arr.let {
-            tickerAdapter.setItems(it)
-        }
+
         return root
     }
 
-    fun selectedMarket(): String{
-        return "BTC"
-    }
-
-    companion object {
-        fun newInstance() = MainFragment()
+    fun selectedMarket(market: String){
+        Toast.makeText(context, market, LENGTH_SHORT).show()
     }
 }
