+++ NackUn/StudyFirst/app/src/main/res/layout/main_fragment.xml
@@ -1,0 +1,115 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout
+xmlns:android="http://schemas.android.com/apk/res/android"
+xmlns:tools="http://schemas.android.com/tools"
+xmlns:app="http://schemas.android.com/apk/res-auto"
+android:layout_width="match_parent" android:layout_height="match_parent"
+android:background="#FEFBF2">
+<TextView
+        android:layout_alignParentLeft="true"
+        android:layout_alignParentTop="true"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/exchange"
+        android:textSize="24dp"
+        android:layout_marginTop="12dp"
+        android:layout_marginLeft="12dp"
+        android:id="@+id/marketTitle"
+        android:textColor="#000000">
+</TextView>
+<LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="horizontal"
+        android:layout_below="@+id/marketTitle"
+        android:id="@+id/marketLike"
+        android:layout_marginTop="12dp">
+    <TextView
+            android:id="@+id/marketKRW"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:text="@string/marketKRW"
+            android:textSize="20dp"
+            android:gravity="center">
+    </TextView>
+    <TextView
+            android:id="@+id/marketBTC"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:text="@string/marketBTC"
+            android:textSize="20dp"
+            android:gravity="center">
+    </TextView>
+    <TextView
+            android:id="@+id/marketETH"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:text="@string/marketETH"
+            android:textSize="20dp"
+            android:gravity="center">
+    </TextView>
+    <TextView
+            android:id="@+id/marketUSDT"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:text="@string/marketUSDT"
+            android:textSize="20dp"
+            android:gravity="center">
+    </TextView>
+</LinearLayout>
+<LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="horizontal"
+        android:layout_marginTop="8dp"
+        android:layout_marginLeft="12dp"
+        android:layout_marginRight="12dp"
+        android:id="@+id/tickerTitle"
+        android:layout_below="@+id/marketLike">
+    <TextView
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:text="@string/coinName"
+            android:textSize="18dp"
+            android:layout_weight="1"/>
+    <TextView
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:text="@string/currentPrice"
+            android:textSize="18dp"
+            android:layout_weight="1"
+            android:gravity="right"/>
+    <TextView
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:text="@string/comparePrice"
+            android:textSize="18dp"
+            android:layout_weight="1"
+            android:gravity="right"/>
+    <TextView
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:text="@string/changePrice"
+            android:textSize="18dp"
+            android:layout_weight="1"
+            android:gravity="right"/>
+
+</LinearLayout>
+<ScrollView
+        android:layout_below="@+id/tickerTitle"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
+
+    <android.support.v7.widget.RecyclerView
+            android:layout_width="match_parent"
+            android:layout_height="match_parent" xmlns:app="http://schemas.android.com/apk/res-auto"
+            android:id="@+id/tickerRecyclerView"
+            app:layoutManager="android.support.v7.widget.LinearLayoutManager"
+            xmlns:android="http://schemas.android.com/apk/res/android">
+    </android.support.v7.widget.RecyclerView>
+</ScrollView>
+</RelativeLayout>
