--- NackUn/StudyFirst/app/src/main/res/layout/main_recyclerview.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.v7.widget.RecyclerView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" xmlns:app="http://schemas.android.com/apk/res-auto"
-        android:id="@+id/tickerRecyclerView"
-        app:layoutManager="android.support.v7.widget.LinearLayoutManager"
-        xmlns:android="http://schemas.android.com/apk/res/android">
-</android.support.v7.widget.RecyclerView>
